# backend/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VitaLink API
  version: 1.0.0
  summary: Versioned API contract for the VitaLink backend
  description: |
    OpenAPI specification for the currently implemented VitaLink backend.

    Notes:
    - The canonical integration base path is `/api/v1`.
    - Legacy `/api/*` routes remain available for backward compatibility but are deprecated.
    - Authentication uses revocable bearer JWT sessions with opaque refresh tokens.
    - Most application responses use `{ statusCode, message, data, success }`; authentication, rate-limit, and direct authorization-guard failures may use the compatible `{ success, message }` subset.
    - Administrative and statistics routes enforce persisted RBAC V2 capabilities. Capability names are namespaced as `platform.*` for global authority and `tenant.*` for one-hospital authority.
    - Administrator roles are fixed: `app_admin` is global and protected, `hospital_admin` is limited to one active hospital, and `auditor` is global and hard read-only. Doctor and Patient authorization is not editable through administrator role policy.
    - Versioned responses expose `X-API-Version` and `X-API-Supported-Versions`. Legacy `/api/*` responses also expose `Deprecation`, `Sunset`, and `Link` headers.
...
    - Multipart upload endpoints accept a single form field named `file`; validation failures such as unsupported MIME type or size limit return 400, while cloud storage failures return 507.
  contact:
...
  license:
    name: MIT
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /admin/access/me:
    get:
      x-admin-self-service: true
      x-admin-scope: self
      x-admin-roles: [app_admin, hospital_admin, auditor]
      tags: [Admin]
      operationId: getCurrentAdminAccess
      summary: Get effective administrator access
      description: Returns the current persisted RBAC V2 snapshot for any authenticated administrator. This self-service read has no capability requirement.
      responses:
        "200":
          description: Effective administrator access retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccessResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /admin/admin-accounts:
    get:
      x-required-capability: platform.admin_accounts.read
      x-admin-scope: global
      x-admin-roles: [app_admin]
      tags: [Admin]
      operationId: listAdminAccounts
      summary: List manageable administrator accounts
      description: Application Admin-only. Returns Hospital Admin and System Auditor accounts; Application Admin accounts are excluded.
      responses:
        "200":
          description: Administrator accounts retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountListResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
    post:
      x-required-capability: platform.admin_accounts.manage
      x-admin-scope: global
      x-admin-roles: [app_admin]
      tags: [Admin]
      operationId: createAdminAccount
      summary: Create a Hospital Admin or System Auditor account
      description: Application Admin-only. hospital_admin requires an active hospital; auditor forbids a hospital; app_admin cannot be created through the portal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAccountCreateRequest"
      responses:
        "201":
          description: Administrator account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountCreateResponse"
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/admin-accounts/{id}:
    put:
      x-required-capability: platform.admin_accounts.manage
      x-admin-scope: global
      x-admin-roles: [app_admin]
      tags: [Admin]
      operationId: updateAdminAccount
      summary: Update an administrator account
      description: Application Admin-only. Role, hospital scope, or state changes bump security_version and revoke sessions. Application Admin targets are denied.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAccountUpdateRequest"
      responses:
        "200":
          description: Administrator account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountUpdateResponse"
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/admin-accounts/{id}/mfa/reset:
    post:
      x-required-capability: platform.admin_accounts.manage
      x-admin-scope: global
      x-admin-roles: [app_admin]
      tags: [Admin]
      operationId: resetAdminAccountMfa
      summary: Reset an administrator authenticator factor
      description: Application Admin-only supervised recovery for an active Hospital Admin or System Auditor. Returns the replacement TOTP setup secret once.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Administrator authenticator reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountMfaResetResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /auth/login`. Send as `Authorization: Bearer <token>`.
        Patient routes require PATIENT users and doctor routes require DOCTOR users.
        Admin/statistics routes additionally resolve the current persisted RBAC V2 policy on every request, enforce a namespaced capability and global/tenant scope, and never authorize from mutable JWT permission claims.
...
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Resource identifier
...
  responses:
...
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not authorized by role, capability, scope, or resource relationship
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ForbiddenErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ApiResponseBase:
      type: object
      required: [statusCode, message, success]
      properties:
        statusCode:
          type: integer
        message:
          type: string
        success:
          type: boolean
        data:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items: {}
            - type: string
            - type: number
            - type: boolean
            - type: "null"
...
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          example:
            statusCode: 401
            message: Invalid credentials
            data: null
            success: false
    ForbiddenErrorResponse:
      type: object
      required: [success, message]
      properties:
        statusCode:
          type: integer
          enum: [403]
          description: Present when emitted through the common error handler.
        message: { type: string }
        data: { type: "null" }
        success: { type: boolean, enum: [false] }
        required_capability:
          allOf:
            - $ref: "#/components/schemas/AdminCapability"
          description: Safe canonical capability when exactly one capability guards the route.
        required_capabilities:
          type: array
          items: { $ref: "#/components/schemas/AdminCapability" }
          description: Safe alternatives when any one of several capabilities can authorize the route.
        policy_version:
          type: integer
          minimum: 1
          description: Safe persisted policy version used for the denied request, when resolved.
      example:
        statusCode: 403
        message: Administrator role does not have the required capability.
        data: null
        success: false
        required_capability: tenant.patients.assign
        policy_version: 7
...
    ValidationErrorItem:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                errors:
                  type: array
                  items:
                    $ref: "#/components/schemas/ValidationErrorItem"
...
    AdminRoleKey:
      type: string
      enum: [app_admin, hospital_admin, auditor]
    EditableAdminRoleKey:
      type: string
      enum: [hospital_admin, auditor]
...
    AdminCapability:
      type: string
      enum:
        - platform.hospitals.read
        - platform.hospitals.manage
        - platform.admin_accounts.read
        - platform.admin_accounts.manage
        - platform.role_policy.read
        - platform.role_policy.manage
        - platform.audit.read
        - platform.analytics.read
        - platform.billing.read
        - platform.billing.manage
        - platform.system_config.read
        - platform.system_config.manage
        - platform.system_health.read
        - platform.notifications.broadcast
        - tenant.dashboard.read
        - tenant.doctors.read
        - tenant.doctors.manage
        - tenant.patients.read
        - tenant.patients.manage
        - tenant.patients.assign
        - tenant.accounts.status.manage
        - tenant.credentials.reset
        - tenant.audit.read
        - tenant.analytics.read
        - tenant.billing.read
        - tenant.billing.checkout
        - tenant.notifications.broadcast
        - tenant.operations_health.read
...
    AdminAccessHospital:
      type: object
      required: [id, code]
      properties:
        id: { type: string }
        code: { type: string }
    AdminAccess:
      type: object
      required: [schema_version, user_id, role, scope, hospital, effective_capabilities, policy_version, read_only]
      properties:
        schema_version: { type: integer, enum: [2] }
        user_id: { type: string }
        role: { $ref: "#/components/schemas/AdminRoleKey" }
        scope: { type: string, enum: [global, tenant] }
        hospital:
          oneOf:
            - $ref: "#/components/schemas/AdminAccessHospital"
            - type: "null"
        effective_capabilities:
          type: array
          uniqueItems: true
          items: { $ref: "#/components/schemas/AdminCapability" }
        policy_version: { type: integer, minimum: 1 }
        read_only:
          type: boolean
          description: Always true for System Auditor.
    AdminAccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/AdminAccess" }
...
    AdminAccountHospital:
      type: object
      required: [id, code]
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
    AdminAccount:
      type: object
      required: [id, login_id, email, name, role, hospital, is_active, mfa_enabled]
      properties:
        id: { type: string }
        login_id: { type: string, format: email }
        loginId:
          type: string
          format: email
          deprecated: true
          description: Legacy camelCase alias; use login_id.
        email: { type: string, format: email }
        name: { type: string }
        role: { $ref: "#/components/schemas/EditableAdminRoleKey" }
        hospital:
          oneOf:
            - $ref: "#/components/schemas/AdminAccountHospital"
            - type: "null"
        is_active: { type: boolean }
        mfa_enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AdminAccountCreateRequest:
      type: object
      required: [name, email, role]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        email:
          type: string
          format: email
          description: Email is the administrator login ID.
        role: { $ref: "#/components/schemas/EditableAdminRoleKey" }
        hospital_id:
          type: string
          description: Required for hospital_admin and forbidden for auditor; must identify an active hospital accepting members.
      additionalProperties: false
    AdminAccountUpdateRequest:
      type: object
      minProperties: 1
      properties:
        role: { $ref: "#/components/schemas/EditableAdminRoleKey" }
        name: { type: string, minLength: 1 }
        hospital_id: { type: string, minLength: 1 }
        hospital:
          type: string
          minLength: 1
          description: Compatibility hospital code/id alias.
        is_active: { type: boolean }
        status:
          type: string
          enum: [active, inactive]
          description: Compatibility alias for is_active; conflicting values are rejected.
      additionalProperties: false
    AdminAccountListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              type: object
              required: [admin_accounts]
              properties:
                admin_accounts:
                  type: array
                  items: { $ref: "#/components/schemas/AdminAccount" }
    AdminAccountCreateResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              type: object
              required: [admin_account, temporary_password, must_change_password]
              properties:
                admin_account: { $ref: "#/components/schemas/AdminAccount" }
                temporary_password:
                  type: string
                  description: One-time generated password returned only in this response.
                must_change_password: { type: boolean, enum: [true] }
    AdminAccountUpdateResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              type: object
              required: [admin_account, invalidated_sessions, revocation_cleanup_completed, security_version_bumped]
              properties:
                admin_account: { $ref: "#/components/schemas/AdminAccount" }
                invalidated_sessions: { type: integer, minimum: 0 }
                revocation_cleanup_completed: { type: boolean }
                security_version_bumped: { type: boolean }
    AdminMfaRecoverySetup:
      type: object
      required: [secret, otpauth_url, challenge_cleanup_completed]
      properties:
        secret:
          type: string
          description: Plaintext replacement TOTP secret returned only in this supervised recovery response.
        otpauth_url: { type: string, format: uri }
        challenge_cleanup_completed: { type: boolean }
    AdminAccountMfaResetResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              type: object
              required: [admin_account, factor_type, setup, invalidated_sessions, revocation_cleanup_completed, challenge_cleanup_completed]
              properties:
                admin_account: { $ref: "#/components/schemas/AdminAccount" }
                factor_type: { type: string, enum: [AUTHENTICATOR_APP] }
                setup: { $ref: "#/components/schemas/AdminMfaRecoverySetup" }
                invalidated_sessions: { type: integer, minimum: 0 }
                revocation_cleanup_completed: { type: boolean }
                challenge_cleanup_completed: { type: boolean }
...